Create java code area
This code defines a class named "Rectangle" with two variables for length and width. Then we have two methods: one for calculating the area (calculateArea) and the other for calculating the perimeter (calculatePerimeter). Finally, in the main method, an instance of this class is created, and the area and perimeter of the rectangle are calculated and printed.